wip: working on examples
