+++ README.md
@@ -4,12 +4,14 @@ Golang colorized JSON output
 `neilotoole/jsoncolor` is a fork of `github.com/segmentio/encoding`'s `json` package that
 provides colorized JSON output.
 
-## Example:
+## Usage
 
 Get the package:
 
 ```shell
-go get github.com/neilotoole/jsoncolor
+go get -u github.com/neilotoole/jsoncolor
 ```
 
+```go
 
+```
